feat: logout, admin-reducer
