+++ src/components/AdminPage/AdminHeader/AdminHeader.scss
@@ -96,6 +96,35 @@
         margin-right: 8px;
       }
     }
+    &__dropdown {
+      position: absolute;
+      top: 46px;
+      right: 5px;
+      width: 105px;
+      height: 36px;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      background-color: #fff;
+      box-shadow: 0px 1px 0px rgba(90, 97, 105, 0.11),
+        0px 2px 4px rgba(90, 97, 105, 0.12), 0px 5px 5px rgba(90, 97, 105, 0.06),
+        0px 3.5px 35px rgba(90, 97, 105, 0.1);
+      border-radius: 3px;
+      z-index: 5;
+    }
+    &__dropdown-btn {
+      font-style: normal;
+      font-weight: normal;
+      font-size: 15px;
+      line-height: 17px;
+      color: #5a6169;
+      background-color: #fff;
+      border: none;
+      cursor: pointer;
+      &:focus {
+        outline: none;
+      }
+    }
     &__dropdown-icon {
       cursor: pointer;
     }
