+++ src/store/redux-store.js
@@ -1,15 +1,17 @@
 import { applyMiddleware, combineReducers, compose, createStore } from "redux"
 import thunkMiddleware from "redux-thunk"
 import orderReducer from "./order-reducer"
+import adminReducer from "./admin-reducer"
 
 const reducers = combineReducers({
   order: orderReducer,
+  admin: adminReducer,
 })
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
 const store = createStore(
   reducers,
-  composeEnhancers(applyMiddleware(thunkMiddleware))
+  composeEnhancers(applyMiddleware(thunkMiddleware)),
 )
 
 export default store
